[Add] 기능 개발
퀘스트 초기화 구현 완료

+++ Growing-a-Cactus/Assets/Scripts/Others/GameManager.cs
@@ -159,7 +159,7 @@ public class GameManager : MonoBehaviour
         UpdateGoldText();
     }
 
-    public void DecreaseGold(int amount)
+    public void DecreaseGold(double amount)
     {
         if (gold >= amount)
         {
